version 2021.12.23-1
complete rewritting of the code without the creation of the index.html file. The needed parameters are located into a "definitions.txt" files which is used with the php GUI file.
Also, the library has been rewritten too with some class effects instead of "classical" functions.
It seems to be more clear.